calendar_assign_project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
firebase-admin==6.4.0
mysqlclient==2.2.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
pytz==2024.1
